#609 Added new API for sending email to student. API automatically detects the current user (who is supposed to be a teacher) and uses that id as the teacher_id. User_id field has to be added to the request.
Currently, the API call is called when a teacher gives feedback on a student submission in the submission view in popup, under 'code'.

Alongside the new API, I also made a controller, service and repository. In case email functionality is needed in the future, we will have a single place to keep all email-related functions.

The current solution does not work in localhost and HAS NOT been tested in development nor live conditions. It is simply a prototype/template to try out by deploying to development. If development has SMPT configured correctly, then in theory all should work.

If for any reasons it does not work, but SMPT configurations are correct, then I suggest trying to create StdClass objects for teacher and student objects. Currently, teacher and student are fetched from DB, they should be objects, but I'm not 100% sure that PHPMailer accepts them.