#609 implemented message API
Added message API to send a notification when teacher comments on student submission. It is also possible to configure email notification in user settings. Email notification can be set as default behaviour as well, but for testing purposes, it is currently disabled by default.

Note: code is not yet finished and does not work 100% as intended. Will update soon.